202505_Audio_Assistant/requirement.txt
langchain-core==0.3.60
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
ollama==0.4.8
openai-whisper==20240930
Pillow==10.4.0
pydantic==2.11.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.11.4
soundfile==0.13.1
streamlit==1.36.0
tiktoken==0.9.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.38.2
TTS==0.22.0
typing-inspect==0.4.0
yarl==1.20.0